service-provider-service/requirements.txt
fastapi
uvicorn
pymongo[srv]
python-dotenv
pydantic
email-validator